requirements.txt
# Core AI and Deep Learning
torch>=1.7.1
torchvision>=0.8.2
torchaudio>=0.7.2
numpy>=1.19.2
scipy>=1.5.2
scikit-image>=0.17.2

# Computer Vision
opencv-python>=4.5.1
Pillow>=8.0.0
lpips>=0.1.4

# Face Recognition and Analysis
insightface>=0.7.3
onnxruntime-gpu>=1.8.0
dlib>=19.22.0

# Web Framework
Flask>=2.0.0
Flask-CORS>=3.0.10
Werkzeug>=2.0.0

# Image and Video Processing
basicsr>=1.4.2
facelib>=0.2.5
ffmpeg-python>=0.2.0

# Utilities
addict>=2.4.0
future>=0.18.2
lmdb>=1.2.1
pyyaml>=5.4.1
requests>=2.25.1
tqdm>=4.60.0
yapf>=0.30.0
gdown>=4.5.1

# Development and Debugging
tb-nightly>=2.4.0

# Additional Dependencies
matplotlib>=3.3.0
seaborn>=0.11.0
pandas>=1.3.0